harden the backend schema, wire contract, security and external adapters
Persistence: the character columns that the relational model leaves unbounded
are mapped as unbounded storage instead of an implicit 255-character bound, so
a full-length post no longer risks truncation; the settings primary key keeps a
bounded type. Identifiers are Long throughout the entities, repositories,
mappers and services. The response-to-tweet association loads lazily and the
paginated read supplies an entity graph, so listing responses no longer issues
one query per parent. The schema is unchanged in shape: four tables, twenty
columns, one ordered association, no added not-null, unique, length or index.

Wire contract: a response representation always carries its stored identifier
while every non-key component is nullable, so a sparse stored row is served
rather than failing; list components are normalised to unmodifiable lists. A
partial response update distinguishes an absent key from an explicit null, so
presence decides what is applied. Both list routes bound the effective page
size while preserving the page and per_page names, their 1 and 10 defaults,
1-based numbering and the four envelope keys.

Analytics and background work: the summary reads its counts from a single
snapshot, so the pending total can never disagree with the approved total. The
unanswered-tweet query returns one bounded, ascending slice instead of the
whole backlog.

External adapters: the language-model client carries an explicit request
timeout and retry limit and is closed on shutdown; the Notion client owns
finite connect and read timeouts; the streaming client targets the current X
API host. Generation rejects an absent, refused or blank completion instead of
storing an empty draft, and returns text so a representation is only ever built
from a stored row. The model call runs outside any database transaction, so a
slow provider cannot hold a pooled connection. No path publishes to X.

Security: tokens are signed and verified with HS256 only, must carry a subject,
issued-at and expiry, and their lifetime is validated at startup within one to
sixty minutes using checked arithmetic. The bcrypt credential hash is validated
for structure and cost when the encoder is created. Unauthenticated requests
receive an empty 401 with a bearer challenge, the framework logout route is
removed, credential fields are length-limited, and neither token material,
principals nor provider exception text reaches the log.

Settings: default rows are seeded once on readiness with insert-only semantics
that tolerate a concurrent writer, so an operator's edit survives a restart.

Error handling: malformed bodies, unsupported methods and unsatisfiable media
types now receive their own safe envelopes, always serialised as JSON, while the
two original error bodies and every per-route message are unchanged.

Tests: the suite covers 408 cases across twelve classes, pinning the physical
column types and lengths, the identifier widths, the association ordering, the
token and credential rules, the seeding semantics, the adapter contracts and
every error envelope through a real dispatch.

Documentation: the decision log records every non-trivial choice with its
alternatives, reasoning and risks, and a bidirectional traceability matrix maps
each retired source construct to its replacement.

Operations: continuous integration builds and verifies with JDK 21 and Maven;
the container image builds in two stages from digest-pinned bases and runs as a
non-root user on port 5000; the image build and deployment script reference the
backend Dockerfile explicitly.

+++ backend/src/main/java/com/codeskeptic/scanner/entity/AiTool.java
@@ -14,9 +14,10 @@ import jakarta.persistence.Table;
  * Schema generation is driven from these annotations (see docs/DECISION_LOG.md DL-026).
  * {@code id} is carried as a {@link String} at the wire boundary (see docs/DECISION_LOG.md DL-023).
  *
- * <p>{@code id} is persisted as {@link Integer}, the type declared at
- * backend/app/db/models.py:L35, and generated DDL therefore declares the column {@code integer} on
- * every supported vendor — DL-070 — see docs/DECISION_LOG.md.
+ * <p>{@code id} is persisted as {@link Long}, the identifier type AAP §0.4.1.4 declares for
+ * {@code AiToolRepository} and the type {@link Tweet} and {@link Response} also carry. TR-3 maps the
+ * source {@code Column(Integer, primary_key=True)} of backend/app/db/models.py:L35 onto
+ * {@code Integer} or {@code Long} — DL-070 — see docs/DECISION_LOG.md.
  *
  * <p>{@code name} and {@code description} declare {@code length = Integer.MAX_VALUE}, which renders
  * each vendor's unbounded character type and reproduces the unbounded {@code Column(String)} at
@@ -30,11 +31,11 @@ import jakarta.persistence.Table;
 public class AiTool {
 
     // backend/app/db/models.py:L35
-    // Integer persistence type — DL-070 — see docs/DECISION_LOG.md
+    // Long persistence type, per AAP §0.4.1.4 and TR-3 — DL-070 — see docs/DECISION_LOG.md
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     // backend/app/db/models.py:L36
     // Unbounded character mapping — DL-068 — see docs/DECISION_LOG.md
@@ -65,11 +66,11 @@ public class AiTool {
         this.description = description;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -111,7 +112,7 @@ public class AiTool {
         if (!(other instanceof AiTool that)) {
             return false;
         }
-        Integer thisId = this.getId();
+        Long thisId = this.getId();
         return thisId != null && thisId.equals(that.getId());
     }
