+++ backend/src/main/java/com/codeskeptic/scanner/repository/AiToolRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Spring Data JPA repository for the {@link AiTool} entity, which maps the {@code ai_tools} table.
  *
- * <p>The identifier type is {@link Integer}, matching the {@code @Id} field of {@link AiTool} — see
- * docs/DECISION_LOG.md DL-070. The {@code ai_tools} table declares no association. No member is
- * declared on this interface; both consumers call the inherited surface:
+ * <p>The identifier type is {@link Long}, matching the {@code @Id} field of {@link AiTool} — see
+ * docs/DECISION_LOG.md DL-070. The {@code ai_tools} table declares no association. Both consumers
+ * call the inherited surface:
  *
  * <ul>
  *   <li>{@code findAll()} returns every {@code ai_tools} row, and {@code TweetStreamClient} takes
@@ -18,12 +18,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
  *       reports it as {@code tracked_ai_tools} — see docs/DECISION_LOG.md DL-041.
  * </ul>
  *
- * <p>Spring Data supplies the implementation as a runtime proxy, registered by the component scan
- * of the application class. Transaction boundaries are declared on the {@code @Service} methods
- * that call this interface. Creation of the {@code ai_tools} table is driven by
- * {@code spring.jpa.hibernate.ddl-auto} from the {@link AiTool} annotations — see
- * docs/DECISION_LOG.md DL-026. This interface declares no JPQL and no SQL — see
- * docs/DECISION_LOG.md DL-027.
+ * <p>Spring Data supplies the implementation as a runtime proxy. Transaction boundaries are declared
+ * on the {@code @Service} methods that call this interface, and the {@code ai_tools} table is created
+ * from the {@link AiTool} annotations by {@code spring.jpa.hibernate.ddl-auto} — see
+ * docs/DECISION_LOG.md DL-026.
  *
  * <p>Usage:
  *
@@ -38,7 +36,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @see AiTool
  */
 // Ported from backend/app/db/database.py:L10-13 (faithful port) — see docs/DECISION_LOG.md
-// Deviation from the literal JpaRepository<AiTool, Long> of AAP §0.4.1.4: the identifier type is
-// Integer, matching the @Id field of entity/AiTool — DL-070 — see docs/DECISION_LOG.md
-public interface AiToolRepository extends JpaRepository<AiTool, Integer> {
+public interface AiToolRepository extends JpaRepository<AiTool, Long> {
 }
